Moves sound directory, renames model classes to include .model
